backend/requirements.txt
django>=5.0,<6.0
channels>=4.0
daphne>=4.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-decouple>=3.8
django-filter>=24.0
Pillow>=10.0
anthropic>=0.40.0
pypdf>=4.0
requests>=2.31.0
# SMS — OTP delivery
twilio>=9.0.0
# Push notifications
firebase-admin>=6.5.0
# DOCX template rendering (jinja2-based merge fields for lease generation)
docxtpl>=0.16.0
# HTML → PDF export (reverse of PDF import)
xhtml2pdf>=0.2.11
# Local vector RAG for contract / policy documents
chromadb>=0.5.0
# High-quality embeddings for RAG (nomic-embed-text-v1.5)
sentence-transformers>=3.0.0
einops>=0.8.0
python-docx>=1.1.0
# MCP server for tenant chat context (used by Cursor / external agents)
fastmcp>=2.0.0
# GraphQL API (co-exists with DRF for flexible lease/template queries)
graphene-django>=3.2.0
google-auth
# Testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
pytest-mock>=3.12
